moves struct funcs out into their own dedicated files to help remember where things are
