+++ snakes/go/pathy-snake/gamestate-funcs.go
@@ -1,0 +1,8 @@
+package main
+
+func (g GameState) wrapped() bool {
+	if g.Game.Ruleset.Name == "wrapped" {
+		return true
+	}
+	return false
+}
