+++ snakes/go/pathy-snake/helpers.go
@@ -64,141 +64,6 @@ func snakeContains(body []Coord, coord Coord) bool {
 	return false
 }
 
-func (c Coord) onLeftEdge() bool {
-	if c.X == 0 {
-		return true
-	}
-	return false
-}
-
-func (c Coord) onRightEdge(width int) bool {
-	if c.X == width-1 {
-		return true
-	}
-	return false
-}
-
-func (c Coord) onTopEdge(height int) bool {
-	if c.Y == height-1 {
-		return true
-	}
-	return false
-}
-
-func (c Coord) onBottomEdge() bool {
-	if c.Y == 0 {
-		return true
-	}
-	return false
-}
-
-func (c Coord) inBottomLeft() bool {
-	if c.X == 0 && c.Y == 0 {
-		return true
-	}
-	return false
-}
-
-func (coord Coord) inBottomRight(width int) bool {
-	if coord.X == width-1 && coord.Y == 0 {
-		return true
-	}
-	return false
-}
-
-func (c Coord) inTopLeft(height int) bool {
-	if c.X == 0 && c.Y == height-1 {
-		return true
-	}
-	return false
-}
-
-func (c Coord) inTopRight(width int, height int) bool {
-	if c.X == width-1 && c.Y == height-1 {
-		return true
-	}
-	return false
-}
-
-func (s Battlesnake) onLeftEdge() bool {
-	if s.Head.X == 0 {
-		return true
-	}
-	return false
-}
-
-func (s Battlesnake) onRightEdge(width int) bool {
-	if s.Head.X == width-1 {
-		return true
-	}
-	return false
-}
-
-func (s Battlesnake) onTopEdge(height int) bool {
-	if s.Head.Y == height-1 {
-		return true
-	}
-	return false
-}
-
-func (s Battlesnake) onBottomEdge() bool {
-	if s.Head.Y == 0 {
-		return true
-	}
-	return false
-}
-
-func (s Battlesnake) inBottomLeft() bool {
-	if s.Head.X == 0 && s.Head.Y == 0 {
-		return true
-	}
-	return false
-}
-
-func (s Battlesnake) inBottomRight(width int) bool {
-	if s.Head.X == width-1 && s.Head.Y == 0 {
-		return true
-	}
-	return false
-}
-
-func (s Battlesnake) inTopLeft(height int) bool {
-	if s.Head.X == 0 && s.Head.Y == height-1 {
-		return true
-	}
-	return false
-}
-
-func (s Battlesnake) inTopRight(width int, height int) bool {
-	if s.Head.X == width-1 && s.Head.Y == height-1 {
-		return true
-	}
-	return false
-}
-
-func (s Battlesnake) ownSnake(state GameState) bool {
-	if s.ID == state.You.ID {
-		return true
-	}
-	return false
-}
-
-func (g GameState) wrapped() bool {
-	if g.Game.Ruleset.Name == "wrapped" {
-		return true
-	}
-	return false
-}
-
-func (c Coord) isNextToSnakeHead(state GameState) bool {
-	return isNextToSnakeHead(c, state)
-}
-
-func (c Coord) Surrounded(state GameState) bool {
-	// log that we are checking if a coord is surrounded.
-	return isSurrounded(c, state)
-}
-
 func isNextToSnakeHead(coord Coord, state GameState) bool {
 	// get the four cells around us.
 	above := Coord{X: coord.X, Y: coord.Y + 1}
@@ -243,38 +108,6 @@ func isNextToSnakeHead(coord Coord, state GameState) bool {
 		}
 	}
 	return false
-}
-
-func (c Coord) cellAbove(height int) Coord {
-	// assumes that every game is wrapped.
-	if c.onTopEdge(height) {
-		return Coord{X: c.X, Y: 0}
-	}
-	return Coord{X: c.X, Y: c.Y + 1}
-}
-
-func (c Coord) cellBelow(height int) Coord {
-	// assumes that every game is wrapped.
-	if c.onBottomEdge() {
-		return Coord{X: c.X, Y: height - 1}
-	}
-	return Coord{X: c.X, Y: c.Y - 1}
-}
-
-func (c Coord) cellLeft(width int) Coord {
-	// assumes that every game is wrapped.
-	if c.onLeftEdge() {
-		return Coord{X: width - 1, Y: c.Y}
-	}
-	return Coord{X: c.X - 1, Y: c.Y}
-}
-
-func (c Coord) cellRight(width int) Coord {
-	// assumes that every game is wrapped.
-	if c.onRightEdge(width) {
-		return Coord{X: 0, Y: c.Y}
-	}
-	return Coord{X: c.X + 1, Y: c.Y}
 }
 
 // function to imulate contains for an array  of coords.
